refactored to websocket_routes and updated Board
