+++ public/js/main.js
@@ -1,9 +1,12 @@
 const $message = $("#message")
+const $task = $("createTask")
 const userName = $("#userName").val()
 const userProject = $("#userProject").val()
 
 $(function() {
   let socket = io()
+
+  //chat path
   $message.submit(res => {
     socket.emit("chat message", {
       message: $("#m").val(),
@@ -17,4 +20,6 @@ $(function() {
   nsp.on("chat message", function(msg) {
     $("#messages").append($("<li>").text(msg.user + ": " + msg.message))
   })
+
+  //task creation
 })
