+++ views/board/mainBoard.handlebars
@@ -1,5 +1,6 @@
-<h3>Project: {{user.project}}</h3>
 
+{{#if user.project}}
+<h3>Project: {{user.project}}</h3>
 <!-- <canvas class="whiteboard"></canvas> -->
 <div class="mainBoard">
 
@@ -13,6 +14,8 @@
       <input type="hidden" name="task[project]" value="{{user.project}}">
       <label for="name">Task Name:</label>
       <input type="text" name="task[name]">
+      <label for="currentUser">Current User:</label>
+      <input type="text" name="task[currentUser]" value="">
       <label for="startDate">Start Date:</label>
       <input type="date" name="task[startDate]" value="">
       <label for="endDate">End Date:</label>
@@ -40,3 +43,5 @@
     </form>
   </div>
 </div>
+
+{{/if}}
